requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
pydantic==2.7.1
python-multipart==0.0.9
cloudinary==1.40.0
PyJWT==2.8.0
dnspython==2.6.1
